+++ index.web.ts
@@ -14,15 +14,15 @@
  * under the License.
  */
 
-export { default as Client } from './src/client';
+export { default as Client } from './src/api/client';
+export * from './src/api/database/types';
+export * from './src/api/engine/types';
+export * from './src/api/oauthClient/types';
+export * from './src/api/query/types';
+export * from './src/api/transaction/types';
+export * from './src/api/user/types';
 export * from './src/credentials';
-export * from './src/database/types';
-export * from './src/engine/types';
 export type { SdkError } from './src/errors';
 export { ApiError, TransactionError } from './src/errors';
-export * from './src/oauthClient/types';
-export * from './src/query/types';
 export * from './src/relationUtils';
-export * from './src/transaction/types';
 export { VERSION } from './src/types';
-export * from './src/user/types';
