+++ src/api/oauthClient/oauthClientApi.test.ts
@@ -16,7 +16,7 @@
 
 import nock from 'nock';
 
-import { baseUrl, getMockConfig } from '../testUtils';
+import { baseUrl, getMockConfig } from '../../testUtils';
 import { OAuthClientApi } from './oauthClientApi';
 import { Permission } from './types';
